requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.10
packaging==25.0
PyMySQL==1.1.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
setuptools==67.6.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
wheel==0.40.0
